go back to movement from attack: use an event, not undo
EDIT:  why did i switch from a move to an event?: Because I wanted to be able to go BACK to my movement stage, and then press UNDO and undo some of my moves.
HOWEVER, even if I used an event, or a move that calls the event, the UNDO button still tries to UNDO that, instead of undoing the last MOVE made. So sad.