glyphs rendering hexagon, some refactors to be done on display
I want to make a Hex component that just take some coordinates, and then spits out a <g> that is positioned where it should be.
This way, we can layer on the different layers of the board: 1. The hex 2: the glyph 3: the unit 4: the unit name text 5: altitude text........... ETC